客户管理模块

+++ jeecg-boot/jeecg-boot-module-user/src/main/java/org/jeecg/modules/user/entity/UserAccount.java
@@ -34,7 +34,7 @@ public class UserAccount implements Serializable {
 	/**会员编号*/
 	@Excel(name = "会员编号", width = 15)
     @ApiModelProperty(value = "会员编号")
-    private java.lang.Integer userId;
+    private java.lang.Long userId;
 	/**用户名*/
 	@Excel(name = "用户名", width = 15)
     @ApiModelProperty(value = "用户名")
